Move listen stuff in libevent engine
